rag-battle/requirements.txt
# requirements.txt
# Core LangChain packages
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-experimental>=0.0.47

# Text splitting (separate package in newer versions)
langchain-text-splitters>=0.0.1

# LangGraph for orchestration
langgraph>=0.0.20

# Database & processing
neo4j>=5.14.0
chromadb>=0.4.22

# PDF processing
pymupdf>=1.23.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# OpenAI
openai>=1.12.0

# Optional but recommended
pandas>=2.0.0
numpy>=1.24.0